auth-service/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.32.0
python-multipart==0.0.18
pydantic==2.10.4
sqlalchemy==2.0.36
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
